ipa-kdb: allow multiple non-password auth types for users
Historically we started to clear all bits in user authentication if
RADIUS proxy link was associated with the user account, meaning none of
other methods are possible to use. While it is possible scenario for
RADIUS case, it is not entirely true: PKINIT is not covered by the logic
of ipadb_get_user_auth(), thus even if RADIUS is enabled, PKINIT would
still be possible. In fact, only methods relying on the password in IPA
LDAP would not be possible to use, like OTP or one factor SPAKE.

Thus, we should probably clear OTP bit rather than removing all but
the RADIUS bit.

Signed-off-by: Alexander Bokovoy <abokovoy@redhat.com>